requirements.txt
absl-py==2.0.0
aiohttp==3.8.5
aiosignal==1.3.1
altgraph==0.17.3
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
build==1.0.3
buildozer==1.5.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.5.7
charset-normalizer==3.1.0
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
cssbeautifier==1.14.8
dataclasses-json==0.6.3
Deprecated==1.2.14
distlib==0.3.7
distro==1.8.0
Django==4.2.2
django-cors-headers==4.1.0
django-decouple==2.1
django-filter==23.3
django-rest-framework==0.1.0
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dm-tree==0.1.8
docutils==0.20.1
docx2pdf==0.1.8
drf-spectacular==0.26.3
drf-spectacular-sidecar==2023.7.1
EditorConfig==0.12.3
Faker==18.10.1
fastapi==0.108.0
filelock==3.12.2
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.12.2
gast==0.5.4
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-generativeai==0.3.2
google-pasta==0.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.20.1
humanfriendly==10.0
idna==3.4
importlib-metadata==6.11.0
importlib-resources==6.1.1
inflection==0.5.1
instaloader==4.10.1
jsbeautifier==1.14.8
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.17.3
keras==2.15.0
kubernetes==28.1.0
langchain==0.0.354
langchain-community==0.0.8
langchain-core==0.1.6
langchain-google-genai==0.0.5
langsmith==0.0.77
libclang==16.0.6
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
ml-dtypes==0.2.0
mmh3==4.0.1
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
namex==0.0.7
numpy==1.26.2
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.3.6
opencv-python==4.8.0.76
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.2
pandas==2.0.3
passlib==1.7.4
pdf2docx==0.5.6
pdfkit==1.0.0
pefile==2023.2.7
pexpect==4.8.0
Pillow==9.5.0
pipdeptree==2.13.0
platformdirs==3.10.0
posthog==3.1.0
proto-plus==1.23.0
protobuf==4.23.4
psycopg2==2.9.9
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pulsar-client==3.4.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.16.1
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.7
PyJWT==2.8.0
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
pypdf==3.17.4
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
sh==2.0.6
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.25
sqlparse==0.4.4
starlette==0.32.0.post1
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tokenizers==0.15.0
toposort==1.10
tqdm==4.65.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.25.0
virtualenv==20.24.3
watchfiles==0.21.0
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
whitenoise==6.6.0
wrapt==1.14.1
yarl==1.9.2
zipp==3.17.0
